pyproject.toml
[tool.poetry]
name = "subd"
version = "0.1.0"
description = ""
authors = ["Moskalenko Maxim <mosmac98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.38"
fastapi = "^0.115.11"
django-environ = "^0.12.0"
psycopg2 = "^2.9.10"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
